+++ src/StateSmithTest/Output/UserConfig/TomlConfigTest.cs
@@ -4,13 +4,17 @@ using FluentAssertions;
 using StateSmith.Input.Settings;
 using StateSmith.Output.UserConfig;
 using StateSmith.Runner;
+using System.Linq;
 using Tomlyn;
 using Tomlyn.Model;
 using Xunit;
 
 namespace StateSmithTest.Output.UserConfig;
 
-public class RenderConfigTomlTest
+/// <summary>
+/// https://github.com/StateSmith/StateSmith/issues/335
+/// </summary>
+public class TomlConfigTest
 {
     /// <summary>
     /// Test out Tomlyn parsing.
@@ -82,13 +86,13 @@ public class RenderConfigTomlTest
     }
 
     [Fact]
-    public void Test1()
+    public void TomlReaderTest()
     {
         RenderConfigAllVars renderConfigAllVars = new();
         RunnerSettings smRunnerSettings = new("");
         var reader = new TomlReader(renderConfigAllVars, smRunnerSettings);
 
-        // $StateSmith(toml)
+        // $CONFIG: toml
         var toml = """"
             [RenderConfig]
             FileTop = "// My file top!"
@@ -123,5 +127,24 @@ public class RenderConfigTomlTest
         smRunnerSettings.transpilerId.Should().Be(TranspilerId.C99);
         smRunnerSettings.algorithmId.Should().Be(AlgorithmId.Balanced1);
         smRunnerSettings.simulation.enableGeneration.Should().BeTrue();
+    }
+
+    [Fact]
+    public void DrawioIntegrationTest()
+    {
+        var (smRunner, files) = TestHelper.CaptureSmRun(diagramPath: "TomlConfig1.drawio");
+        smRunner.Settings.outputStateSmithVersionInfo.Should().BeFalse();
+
+        var hhFile = files.GetCapturesForFileName("TomlConfigEx.hh").Single();
+        var incFile = files.GetCapturesForFileName("TomlConfigEx.inc").Single();
+
+        hhFile.code.Should().Contain("// TEST-FILE-TOP");
+        incFile.code.Should().Contain("// TEST-FILE-TOP");
+
+        hhFile.code.Should().Contain("// TEST-H-FILE-TOP");
+        incFile.code.Should().Contain("// TEST-C-FILE-TOP");
+
+        incFile.code.Should().NotContain("// TEST-H-FILE-TOP");
+        hhFile.code.Should().NotContain("// TEST-C-FILE-TOP");
     }
 }